requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
dj-database-url==0.5.0
Django==3.2.8
django-bootstrap4==3.0.1
django-stdimage==5.3.0
gunicorn==20.1.0
importlib-metadata==2.1.1
mysqlclient==2.0.3
Pillow==8.3.2
psycopg2-binary==2.9.1
PyMySQL==1.0.2
pytz==2021.3
soupsieve==2.2.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
whitenoise==5.3.0
zipp==3.6.0
